+++ CHANGELOG.md
@@ -9,11 +9,15 @@ All notable changes to this project will be documented in this file. Guppy adher
 #### Added
 
 * Support for dynamic font sizes
-* Support for mulitple window scenes
 
 #### Changed
 
 * Minimum iOS version supported is now iOS 13.0
+* Improved contrast of success status in dark mode
+
+#### Fixed
+
+* Fixed issue where search bar text was not readable in dark mode
 
 ---